# Improve PrintableRuneWidth performance.