OLD:
```hs
Benchmark_PrintableRuneWidth-8           4717250               255 ns/op               0 B/op          0 allocs/op
Benchmark_PrintableRuneWidth-8           4819238               257 ns/op               0 B/op          0 allocs/op
Benchmark_PrintableRuneWidth-8           4755002               257 ns/op               0 B/op          0 allocs/op
Benchmark_PrintableRuneWidth-8           4683806               255 ns/op               0 B/op          0 allocs/op
```

NEW:
```hs
Benchmark_PrintableRuneWidth-8           6802281               177 ns/op               0 B/op          0 allocs/op
Benchmark_PrintableRuneWidth-8           7094738               176 ns/op               0 B/op          0 allocs/op
Benchmark_PrintableRuneWidth-8           6692834               177 ns/op               0 B/op          0 allocs/op
Benchmark_PrintableRuneWidth-8           6692290               177 ns/op               0 B/op          0 allocs/op
```